+++ Sphinx/ChangeLog.rst
@@ -3,6 +3,7 @@ ChangeLog
 * July 24, 2014
 
   - Implemented ``width`` parameter in ``add_label()``.
+  - Added ``test/test_button_label.py``.
 
 * 01.06.02 --- July 18, 2014