+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,6 @@
 ChangeLog
 =========
-* July 24, 2014
+* 01.06.03 --- July 24, 2014
 
   - Implemented ``width`` parameter in ``add_label()``.
   - Added ``test/test_button_label.py``.
@@ -11,7 +11,7 @@ ChangeLog
 
 * 01.06.01 --- July 17, 2014
 
-  - Added width parameter in ``add_label()``.
+  - Added (fake) width parameter in ``add_label()``.
   - Corrected gz archive of HTML offline documentation.
 
   - Added private members in all documentation.